+++ packages/apollo-graphql/src/schema/transformSchema.ts
@@ -19,7 +19,8 @@ import {
   GraphQLUnionType,
   isInputObjectType,
   GraphQLInputObjectType,
-  GraphQLInputFieldConfigMap
+  GraphQLInputFieldConfigMap,
+  GraphQLDirective
 } from "graphql";
 import { mapValues } from "../utilities/mapValues";
 
@@ -53,7 +54,8 @@ export function transformSchema(
     types: Object.values(typeMap),
     query: replaceMaybeType(schemaConfig.query),
     mutation: replaceMaybeType(schemaConfig.mutation),
-    subscription: replaceMaybeType(schemaConfig.subscription)
+    subscription: replaceMaybeType(schemaConfig.subscription),
+    directives: replaceDirectives(schemaConfig.directives)
   });
 
   function recreateNamedType(type: GraphQLNamedType): GraphQLNamedType {
@@ -144,5 +146,15 @@ export function transformSchema(
       ...arg,
       type: replaceType(arg.type)
     }));
+  }
+
+  function replaceDirectives(directives: GraphQLDirective[]) {
+    return directives.map(directive => {
+      const config = directive.toConfig();
+      return new GraphQLDirective({
+        ...config,
+        args: replaceArgs(config.args)
+      });
+    });
   }
 }